refactor: group struct panel modules
